requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
keyboard==0.13.5
langchain==0.2.0
openai==0.28
selenium==4.21.0
weaviate_client==3.22.1
tiktoken==0.7.0
langchain-community==0.2.0
sentence-transformers==2.7.0
flask_limiter==3.7.0
PyJWT==2.8.0
flask_restx==1.3.0
python-dateutil
redis==5.0.8
flask-httpauth==4.8.0
